Enhancement in QuizzView
The main views in the 'QuizzView' file were refactored to facilitate the management of each one's code. Additionally, images used in the repository's ReadME file were added.
